Remove race condition in HTTPHandler
The HTTPHandler downloads the file requested and stores it in
<cache_dir>/<base64 filename>.
If the TFTP serves two requests simultaneously asking for the same file,
the file will be downloaded twice at the same location and will be
delivered corrupted to the client.
To fix the issue, we make the local filename unique, by appending the
client's IP and the local time to it.