Remove use of ThreadingMixIn
Creating a thread for each request didn't boot performances because to
get a file, you need to make (size of file / blksize) requests which
is... quite a lot.

The drawback of not using threads comes with http handler. When the
client requests a file, we download it locally. During this time, the
other clients are blocked. This is not really a problem if all clients
are legitimate, but we should consider adding the download to
Socketserver's event loop.